test(emitter): shard printer tests by feature
